+++ HISTORY.rst
@@ -120,7 +120,7 @@
   version that does filtering.
 * Update OCR links activation.
 * Update document parsing link activation.
-
+* Add favorite document views tests.
 3.1.11 (2019-04-XX)
 ===================
 * Fix multiple tag selection wizard step.